ovs-dpdk: support late initialization since ovs 2.6.0
OVS 2.6.0 introduces new mechanisms to configure the DPDK netdev
provider. It now relies on the database and allows dynamic runtime
configuration. Network interface binding is more fine grained by
specifying the NIC PCI address and not relying on special port naming
and indexing.

Here we introduce the support of post 2.6.0 OVS-DPDK and change the
relevant tests.

Signed-off-by: Nikolay Nikolaev <nikolay.nikolaev@canonical.com>