ovs-dpdk: allow dpdk-driver value 'none'
Some NICs do not work with vfi-pci or uio_pci_generic. E.g. mxl4/mlx5
which uses relies on the OFED or Kernel drivers (post 4.14).
In this cases we don't want to generate entries in /etc/dpdk/inerfaces.

Here we introduce a new value for 'dpdk-driver' configuration 'none'. If
used the charm will omit adding entries in the aforementioned file.

Signed-off-by: Nikolay Nikolaev <nikolay.nikolaev@canonical.com>