Add bond-mappings configuration option
The current charm does not support creating and managing bonded network
interfaces. They are managed externaly. This is not possible when DPDK
is enabled. In this case OVS exposes the DPDK bond PMD which enslaves
the corresponding attached bond interfaces.

The new bond-mappings configuration option allows such configuration
where mac:bond is specified. When the data-port configuration is processed
the new bond-mappings are consulted to identify if the port belongs to
a bond. If this is true - then the bond is created with the mac designated
interface and the bond is added to the bridge. Subsequently more interfaces
can be added to the same bond.

Signed-off-by: Nikolay Nikolaev <nikolay.nikolaev@canonical.com>