Adds Neutron OVS agent config to DHCP daemon
The neutron-dhcp-agent might rely on options defined in
openvswitch_agent.ini. By default this config file is not passed to
neutron-dhcp-agent daemon, and therefore those options are not
loaded and fall back for the default values and by that inhibits
the desired behavior.

Charm helpers' PR: https://github.com/juju/charm-helpers/pull/615

Depends-on: I39024855c3e42ee135b6ad5e7618a770219b6994

Closes-bug: #1832021
Change-Id: I134c8077ee52ccdb4e383109ecbea27ed1633fb8