Update to classic charms to build using charmcraft in CI
This is for lastest/edge and yoga/edge.

This update is to ensure that the Zuul Canonical CI builds the charm
before functional tests and ensure that that artifact is used for the
functional tests.  This is to try to ensure that the charm that gets
landed to the charmhub is the same charm that was tested with.

Change-Id: Ia56c171397ecf9fa638405fd960593cd57b87047
Co-authored-by: Aurelien Lourot <aurelien.lourot@canonical.com>