Add docs key and point at Discourse
Add the 'docs' key and point it at a Discourse topic
previously populated with the charm's README contents.

When the new charm revision is released to the Charmhub,
this Discourse-based content will be displayed there. In
the absense of the this new key, the Charmhub's default
behaviour is to display the value of the charm's
'description' key.

Change-Id: I85703b441023d58dee7fc2918cad2bc32899ae91
(cherry picked from commit 6e5cf7da7f3f5c1c8f98c6af087f16ae779d5381)