Use charm specific job templates
The openstack-python3-ussuri-jobs template is running on noble (python
3.12) making the pep8 tox target fail due to the dependencies and
versions locked down by the charm.

This change makes the charm to use the
openstack-python3-charm-jobs-ussuri template which runs on focal and
it's owned by the openstack charms team.

For the coverage report, the openstack-cover-charm-jobs template is used
now which provides the same guarantees described earlier

Also included the following change from Marcin Wilk in
order to make the gate green.

Fix Python 3.12 unittest compatibility

Python 3.12 has removed long-deprecated unittest features [1],
therefore some of the existing unit test fail when run on Python
3.12. This patch replaces the 'assertEquals' with 'assertTrue' or
'assertFalse', depending on the usage.

This patch also fixes the 'test_amqp_changed' and 'test_amqp_departed'
test methods which improperly called 'called_with' which is no longer
a valid assertion method [2]. And in fact on Python < 3.12 the
'test_amqp_changed' and 'test_amqp_departed' pass their assertions,
when they actually should fail due to the out-of-sync with
'amqp-relation-changed' and 'amqp-relation-departed' hook code.
In this case the 'called_with' is replaced with 'assert_called_once'.

[1] https://docs.python.org/3.12/whatsnew/3.12.html#id3
[2] https://github.com/python/cpython/issues/100690

Change-Id: I39f53313fb96c48ae9ac99173a8600082c887dcb