Set neutron live_migration_events to True
When this option is set to True, during the live migration, the OVS
agent will only send the "vif-plugged-event" when the destination
host interface is bound. This option also disables any other agent
(like DHCP) to send to Nova this event when the port is provisioned.

NOTE: we are only making this change to the Yoga release of the
      charm since newer releases have removed the config option and
      it now defaults to True.

Closes-bug: #1901707

Change-Id: I802d98862e23b32944e1710fd9d2efe1af3c3742
Signed-off-by: Alan Baghumian <alan.baghumian@canonical.com>